New slices in utils package
