+++ pkg/utils/maths.go
@@ -83,37 +83,3 @@ func PowOf2(exp uint64) (r uint64) {
 	}
 	return r
 }
-
-// @ preserves acc(r_in, 1/2)
-// @ ensures acc(r_out)
-// @ ensures len(r_in) == len(r_out)
-// @ ensures forall i int :: 0 <= i && i < len(r_in) ==> r_in[i] - 1 == r_out[i]
-// @ decreases
-func Decrement(r_in []uint64) (r_out []uint64) {
-	r_out = make([]uint64, len(r_in))
-	// @ invariant acc(r_in, 1/2)
-	// @ invariant acc(r_out)
-	// @ invariant 0 <= i && i <= len(r_in) && i <= len(r_out)
-	// @ invariant forall j int :: 0 <= j && j < i ==> r_in[j] - 1 == r_out[j]
-	// @ decreases len(r_in) - i
-	for i := 0; i < len(r_in); i++ {
-		r_out[i] = r_in[i] - 1
-	}
-	return r_out
-}
-
-// @ preserves acc(r_in, 1/2)
-// @ ensures acc(r_out)
-// @ ensures len(r_in) == len(r_out)
-// @ ensures forall i int :: 0 <= i && i < len(r_in) ==> r_in[i] + add == r_out[i]
-func Increment(r_in []uint64, add uint64) (r_out []uint64) {
-	r_out = make([]uint64, len(r_in))
-	// @ invariant acc(r_in, 1/2)
-	// @ invariant acc(r_out)
-	// @ invariant 0 <= i && i <= len(r_in) && i <= len(r_out)
-	// @ invariant forall j int :: 0 <= j && j < i ==> r_in[j] + add == r_out[j]
-	for i := 0; i < len(r_in); i++ {
-		r_out[i] = r_in[i] + add
-	}
-	return r_out
-}
